requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
speedtest-cli>=2.1.3
pyinstaller>=6.0.0
